TextFromImage/requirements.txt
packaging==24.0
pillow==10.3.0
pytesseract==0.3.10
anyio==4.4.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mutagen==1.47.0
packaging==24.0
pillow==10.3.0
pycryptodomex==3.20.0
pylast==5.3.0
pyperclip==1.8.2
pytesseract==0.3.10
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.1
websockets==12.0
yewtube==2.10.4
youtube-search-python==1.6.6
yt-dlp==2024.5.27
